+++ lightx2v/models/networks/wan/infer/transformer_infer.py
@@ -172,6 +172,106 @@ class BaseWanTransformerInfer(BaseTransformerInfer):
     def switch_status(self):
         self.infer_conditional = not self.infer_conditional
 
+    # only in Wan2.1 TeaCaching
+    def set_attributes_by_task_and_model(self):
+        if self.config.task == "i2v":
+            if self.use_ret_steps:
+                if self.config.target_width == 480 or self.config.target_height == 480:
+                    self.coefficients = [
+                        2.57151496e05,
+                        -3.54229917e04,
+                        1.40286849e03,
+                        -1.35890334e01,
+                        1.32517977e-01,
+                    ]
+                if self.config.target_width == 720 or self.config.target_height == 720:
+                    self.coefficients = [
+                        8.10705460e03,
+                        2.13393892e03,
+                        -3.72934672e02,
+                        1.66203073e01,
+                        -4.17769401e-02,
+                    ]
+                self.ret_steps = 5 * 2
+                self.cutoff_steps = self.config.infer_steps * 2
+            else:
+                if self.config.target_width == 480 or self.config.target_height == 480:
+                    self.coefficients = [
+                        -3.02331670e02,
+                        2.23948934e02,
+                        -5.25463970e01,
+                        5.87348440e00,
+                        -2.01973289e-01,
+                    ]
+                if self.config.target_width == 720 or self.config.target_height == 720:
+                    self.coefficients = [
+                        -114.36346466,
+                        65.26524496,
+                        -18.82220707,
+                        4.91518089,
+                        -0.23412683,
+                    ]
+                self.ret_steps = 1 * 2
+                self.cutoff_steps = self.config.infer_steps * 2 - 2
+
+        elif self.config.task == "t2v":
+            if self.use_ret_steps:
+                if "1.3B" in self.config.model_path:
+                    self.coefficients = [-5.21862437e04, 9.23041404e03, -5.28275948e02, 1.36987616e01, -4.99875664e-02]
+                if "14B" in self.config.model_path:
+                    self.coefficients = [-3.03318725e05, 4.90537029e04, -2.65530556e03, 5.87365115e01, -3.15583525e-01]
+                self.ret_steps = 5 * 2
+                self.cutoff_steps = self.config.infer_steps * 2
+            else:
+                if "1.3B" in self.config.model_path:
+                    self.coefficients = [2.39676752e03, -1.31110545e03, 2.01331979e02, -8.29855975e00, 1.37887774e-01]
+                if "14B" in self.config.model_path:
+                    self.coefficients = [-5784.54975374, 5449.50911966, -1811.16591783, 256.27178429, -13.02252404]
+                self.ret_steps = 1 * 2
+                self.cutoff_steps = self.config.infer_steps * 2 - 2
+
+    # calculate should_calc
+    def calculate_should_calc(self, weights, embed, grid_sizes, x, embed0, seq_lens, freqs, context):
+        # 1. 时间步嵌入调制
+        modulated_inp = embed0 if self.use_ret_steps else embed
+
+        # 2. L1距离计算
+        should_calc = False
+        if self.infer_conditional:
+            if self.cnt < self.ret_steps or self.cnt >= self.cutoff_steps:
+                should_calc = True
+                self.accumulated_rel_l1_distance_even = 0
+            else:
+                rescale_func = np.poly1d(self.coefficients)
+                self.accumulated_rel_l1_distance_even += rescale_func(
+                    ((modulated_inp - self.previous_e0_even).abs().mean() / self.previous_e0_even.abs().mean()).cpu().item()
+                )
+                if self.accumulated_rel_l1_distance_even < self.teacache_thresh:
+                    should_calc = False
+                else:
+                    should_calc = True
+                    self.accumulated_rel_l1_distance_even = 0
+            self.previous_e0_even = modulated_inp.clone()
+
+        else:
+            if self.cnt < self.ret_steps or self.cnt >= self.cutoff_steps:
+                should_calc = True
+                self.accumulated_rel_l1_distance_odd = 0
+            else:
+                rescale_func = np.poly1d(self.coefficients)
+                self.accumulated_rel_l1_distance_odd += rescale_func(
+                    ((modulated_inp - self.previous_e0_odd).abs().mean() / self.previous_e0_odd.abs().mean()).cpu().item()
+                )
+                if self.accumulated_rel_l1_distance_odd < self.teacache_thresh:
+                    should_calc = False
+                else:
+                    should_calc = True
+                    self.accumulated_rel_l1_distance_odd = 0
+            self.previous_e0_odd = modulated_inp.clone()
+
+        # 3. 返回判断
+        return should_calc
+
     # 1. get taylor step_diff when there is two caching_records in scheduler
     def get_taylor_step_diff(self):
         step_diff = 0
